Add ThatExcludeEmptyResponse option
This makes it so that an empty response, after replacing <oob> elements, will not change the path used for matching the <that> pattern in subsequent requests.
This allows the <that> pattern to actually match 'the bot's last utterance' without asynchronous out-of-band requests that produce empty responses (and this no 'utterance') from interfering with the conversation. This is non-standard behaviour according to the specification and Pandorabots's implementation, so it is disabled by default.